Admin dashboard && save configuration

+++ app/Http/Controllers/MetaAPI/MetaController.php
@@ -4,79 +4,117 @@ namespace App\Http\Controllers\MetaAPI;
 
 use App\Http\Controllers\Controller;
 use App\Models\MetaApiConfiguration;
+use App\Traits\ResponsesJSON;
+use App\Traits\UtilityMethods;
 use Illuminate\Http\Client\ConnectionException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Http;
 
 class MetaController extends Controller
 {
+    use ResponsesJSON, UtilityMethods;
+
+    public function __construct()
+    {
+        $this->middleware('meta-api')->except('get');
+        $this->middleware('token')->only('get');
+    }
 
     public function save(Request $request)
     {
 
-        // Controllo se esistono le proprietà nelle configurazioni
-        $conf = $request->configuration;
+        if($request->header('Content-Type') === 'form-data'){
+            $validator = $request->validate([
+                'configuration_file' => 'required|max:4096'
+            ]);
+
+            $file = $request->file('configuration_file');
+            if ($file->getClientMimeType() !== 'application/json') {
+                return redirect()->back()->withErrors(['file_error' => 'The file must be a JSON file']);
+            }
+
+            $file_content = file_get_contents($file);
+            $conf = json_decode($file_content, true);
+
+        } elseif($request->header('Content-Type') === "application/json"){
+            // Controllo se esistono le proprietà nelle configurazioni
+            $conf = $request->all();
+        }
+
+
+        if (!is_array($conf)) {
+            return $this->ResponseError(400, 'Bad request', 'Parsing error');
+        }
+
         if (!key_exists('group', $conf)) {
-            return "missing group";
+            return $this->ResponseError(400, 'Bad request', 'Missing group');
         } elseif (!key_exists('service', $conf)) {
-            return "missing service";
+            return $this->ResponseError(400, 'Bad request', 'Missing service');
         } elseif (!key_exists('operations', $conf)) {
-            return "missing operations";
+            return $this->ResponseError(400, 'Bad request', 'Missing operations');
         }
 
         // Controllo le proprieta di ogni source
         foreach ($conf['operations'] as $key => $operation) {
             if (!key_exists('sources', $conf['operations'][$key]) || !key_exists('result', $conf['operations'][$key])) {
-                return "missing parameters at " . $key;
+                return $this->ResponseError(400, 'Bad request', "Missing parameters at {$key}");
             }
-            foreach ($operation['sources'] as $source) {
+            foreach ($operation['sources'] as $key => $source) {
                 if (!filter_var($source['urlTemplate'], FILTER_VALIDATE_URL)) {
-                    return "invalid source url";
+                    return $this->ResponseError(400, 'Bad request', "Invalid source url at {$key}");
                 }
 
                 if (!is_string($source['description'])) {
-                    return "description must be a string";
+                    return $this->ResponseError(400, 'Bad request', "Description must be a string at {$key}");
                 } elseif (!filter_var($source['required'], FILTER_VALIDATE_BOOL)) {
-                    return "source required must be a boolean";
+                    return $this->ResponseError(400, 'Bad request', "required must be a boolean at {$key}");
                 }
             }
-
         }
-
         $group = $conf['group'];
         $service = $conf['service'];
-        $operations = $conf['operations'];
 
         $obj = MetaApiConfiguration::where('configuration->group', $group)->where('configuration->service', $service)->first();
         if ($obj) {
-            return "This conf already exists";
+            return $this->ResponseError(400, 'Bad request', 'This configuration already exists');
         }
-
-        $routeFile = fopen(base_path() . '\routes\meta-api.php', "a");
-        foreach ($operations as $key => $operation) {
-            $newRoute = "Route::get('/services/" . $group . '/' . $service . '/' . $key . "','MetaController@get');\n";
-            fwrite($routeFile, $newRoute);
-        }
-        fclose($routeFile);
 
         $obj = MetaApiConfiguration::create([
             'configuration' => json_encode($conf),
         ]);
 
+        return response()->json(['data' => $obj],201,['Content-Type' => 'application/json']);
+
     }
 
-    public function get(Request $request)
+    public function get(Request $request, $group, $service, $operation)
     {
-        $path_array = explode('/', $request->path());
-        $obj = MetaApiConfiguration::where('configuration->group', $path_array[2])->where('configuration->service', $path_array[3])->first();
+//        $path_array = explode('/', $request->path());
+        if (preg_match("/[^a-zA-Z]+/", $group)) {
+            return $this->ResponseError(400, 'Bad request', 'Group is invalid');
+        } elseif (preg_match("/[^a-zA-Z]+/", $service)) {
+            return $this->ResponseError(400, 'Bad request', 'Service is invalid');
+        } elseif (preg_match("/[^a-zA-Z]+/", $operation)) {
+            return $this->ResponseError(400, 'Bad request', 'Operation is invalid');
+        }
+
+        $obj = MetaApiConfiguration::where('configuration->group', $group)->where('configuration->service', $service)->first();
+        if (!$obj) {
+            return $this->ResponseError(404, 'Not found', 'Configuration not found');
+        }
+
         $conf = json_decode($obj->configuration, true);
 
+        if (!key_exists($operation, $conf['operations'])) {
+            return $this->ResponseError(404, 'Not found', 'Operation not found');
+        }
+
         // Controllo se la configurazione ha parametri
-        if (key_exists('params', $conf['operations'][$path_array[4]])) {
+        if (key_exists('params', $conf['operations'][$operation])) {
             // Controllo se c'è almeno un parametro dentro params
-            if (count($conf['operations'][$path_array[4]]['params']) > 0) {
+            if (count($conf['operations'][$operation]['params']) > 0) {
                 $query_params = $request->query(); // Query params
-                $params = $conf['operations'][$path_array[4]]['params']; // Parametri della configurazione
+                $params = $conf['operations'][$operation]['params']; // Parametri della configurazione
                 foreach ($params as $key => $param) {
                     // Controllo se i parametri required sono nella richiesta
                     if ($param['required']) {
@@ -101,7 +139,7 @@ class MetaController extends Controller
             }
         }
         // process source
-        $sources = $conf['operations'][$path_array[4]]['sources'];
+        $sources = $conf['operations'][$operation]['sources'];
         $results = [];
         foreach ($sources as $key => $source) {
 
@@ -129,10 +167,15 @@ class MetaController extends Controller
         ]);
     }
 
-    public function delete(MetaApiConfiguration $configuration)
+    public function delete($id)
     {
+        $configuration = MetaApiConfiguration::find($id);
+        if (!$configuration) {
+            return $this->ResponseError(404, 'Not found', 'Configuration not found');
+        }
+
         $configuration->delete();
-        return redirect()->route('admin.configuration.all');
+        return response()->json([], 204);
     }
 
     public function validateType($type, $value)
